control bar and chart setup
