+++ src/components/chartView.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+function ForecastChart({}){
+
+    return(<></>)
+}
+
+export default ForecastChart;
